BAH-4732 | Fix the payload to include modified ones alone and send PUT for parentObs
